Game1.cs creates the eventHandler:
dogImage.MouseDown += new EventHandler(dogImage_MouseDown);

XAMLiteControl.cs looks to see if there has been a LeftButton press and sets a flag in its Update() method.

XAMLiteImage.cs has an Eventhandler called MouseDown.  In its Update() method, it checks whether the flag was set in XAMLiteControl.cs and if there was a collision.  If so, it fires the MouseDown event and the text changes.

git-svn-id: https://xamlite.svn.codeplex.com/svn@93476 81e6e737-5a20-4d0f-a97c-cb01297b46e0