Grid is at a good resting point.  Did a lot of testing to ensure that the different XAMLiteControls would adhere to it.
The only problem area is the label.  It is fine in every way except for when a label is set to HorizontalAlignment.Center and the position of the grid is not in the center.  Since LoadContent has not been called yet for the label, the width of the string has not been measured yet and this causes the draw to start from the center point.  This can be corrected with margins.  Label works in right left positions any where on the grid.

git-svn-id: https://xamlite.svn.codeplex.com/svn@93579 81e6e737-5a20-4d0f-a97c-cb01297b46e0