Simplification and refinement of inheritance for XAMLite continues.
For now, creating new classes to eventually replace the old classes.  These classes have the same name but include the word "New" at the end until the process is complete.

Hierarchy of refined classes thus far:

                                          XAMLiteBaseControl 
                                         /        |         \
                                        /         |          \
                 XAMLiteBaseContentControl   XAMLiteGridNew   XAMLiteImageNew
                 /                \
                /                  \
    XAMLiteBaseText            XAMLiteBaseContent ---
                              /       |               \
                             /        |                \
                 XAMLiteLabelNew  XAMLiteCheckBoxNew  XAMLiteButtonNew

Classes extending from XAMLiteBaseContent have the variable, Content, which is of type object.  This allows the developer to insert either a string or a Date/Time object, etc.

Classes extending from XAMLiteBaseText have the variable, Text, which is of type string only.

Simplification of how a grid sets up its children's positions in this update.

Changed the Position property in XAMLiteBaseControl to protected instead of public since this doesn't exist in WPF.

git-svn-id: https://xamlite.svn.codeplex.com/svn@100592 81e6e737-5a20-4d0f-a97c-cb01297b46e0