Moved the property, Background, to the XAMLiteBaseControl class.
For XAMLiteLabelNew class, created a workaround for measuring the height of the label.  In XNA, this measurement isn't accurate as it includes the line spacing.  This was to allow a background color to be drawn on the control to more accurately hug the shape of the text.

Working on the XAMLiteButtonNew class, so that there is a default button composed of resizeable image and some Content.

git-svn-id: https://xamlite.svn.codeplex.com/svn@100598 81e6e737-5a20-4d0f-a97c-cb01297b46e0