More modifications to the new classes.
The XAMLiteGridNEw class, has a simpler means of placing its children within its boundaries.

The XAMLiteCheckBoxNew class now uses other XAMLiteclasses to create its components.  It now consists of a XAMLiteGridNew, XAMLiteLabelNew, and 4 XAMLiteImageNew classes.  Before, the wheel was reinvented entirely for all of its internal components.

The object of the newer XAMLite is to reuse other XAMLite objects as much as possible when creating new objects.

git-svn-id: https://xamlite.svn.codeplex.com/svn@100601 81e6e737-5a20-4d0f-a97c-cb01297b46e0