More work on positioning of controls and positioning of controls within grids.  Major simplification.  Now all positioning takes place in the base class, rather than the base class and the grid class when it is a child.
This is working well for controls that are not complex (XAMLiteImageNew, XAMLiteLabelNew, XAMLiteButtonNew).  Controls that are made up of a grid with child XAMLite components, such as the XAMLiteCheckBoxNew and XAMLiteTextBoxNew classes are not working correctly yet.  Basically, any grid that contains another grid with controls is still under construction.  

The end result is to be able to take any set of simple XAMLite classes, add them to a XAMLiteGrid, and then end up with any complex component.  For example, a XAMLiteSlider would consist of a XAMLiteGrid (container), XAMLiteRectangle (to denote the range a slider could be moved), and a XAMLiteImage (or another XAMLiteRectangle to represent the slider).

This flexibility would then allow developers to quickly piece together any custom component not yet built.  

git-svn-id: https://xamlite.svn.codeplex.com/svn@100636 81e6e737-5a20-4d0f-a97c-cb01297b46e0