Implemented a Cursor class for XAMLite that allows other Cursor types such as Hand and IBeam.
When Game.IsMouseVisible = true && !Cursors.Arrow, the XAMLiteCursor will kick in.  The cursor class is simply an extension of the XAMLiteImage class and allows the user to set a cursor type for various controls using the XAMLite Cursors enum.  Currently, there is simply the Cursors.Arrow, Cursors.Hand, and Cursors.IBeam.  When Cursors.Arrow is set, the XAMLite cursor hides and the standard game mouse cursor kicks in.

The idea is to override the need of including the dll file for System.Windows.Forms.

git-svn-id: https://xamlite.svn.codeplex.com/svn@100762 81e6e737-5a20-4d0f-a97c-cb01297b46e0