Added an internal variable called, Index, that is set by a grid when it adds Children.  Mostly not being used yet but may come in handy, especially for modifying the visibility of a child of a grid when the grid is Hidden.
Added a Children class that inherits from List<XAMLiteBaseControl> and has an Add(XAMLiteBaseControl) and RemoveAt(index) that overrides the base class so that it can internally set the index, set the child's parent, etc., yet appear to be a normal List function as would be expected in WPF code behind.

To the XAMLiteGrid class, there is now an internal function that gets called by the Children class method, RemoveAt(index) and removes elements from the lists of child opacity and child visibilities.

git-svn-id: https://xamlite.svn.codeplex.com/svn@100807 81e6e737-5a20-4d0f-a97c-cb01297b46e0