ListBox now manages a list of ListBoxItems correctly.  It supports a singly selected ListBoxItem for now.  Support for multi-selected items will come at some future point when time permits.
When an item is selected, it changes to its selected color.  When another item is selected, the original is deselected and the new item has the selected color.

If a place outside of the ListBox is selected, a selected item in the ListBox will changed to an unfocused color.  When reselected, the focused selected color returns.

git-svn-id: https://xamlite.svn.codeplex.com/svn@100841 81e6e737-5a20-4d0f-a97c-cb01297b46e0