The BXAMLiteBaseControl class now has an Event called LostFocus, the same as WPF, and is fired off when a control loses its focus, if it is Focusable.
Also, the ComboBox now closes when a MouseDown occurs outside of itself.  It also closes and opens as the top portion is clicked.

Now just to add the button portion of the control on the right and fix some minor width adjustment issues based on the text contained and the new control will be completed.

git-svn-id: https://xamlite.svn.codeplex.com/svn@100867 81e6e737-5a20-4d0f-a97c-cb01297b46e0