A XAMLiteGrid can now have additional Children added at runtime.  Before, the grid's children had to be set during the initial load content and could not handle additional children being added later.
The XAMLiteListBox and XAMLiteComboBox may now also have Items added at runtime.  However, at this point, it must be an entire set, not an additional item added to an already built set of Items. For now, remove the current Items and then replace with the new set.  There are some bugs when one adds additional children to an already existing list.

The XAMLiteComboBox now has the property, IsEditable.  When IsEditable is true, the contents of the text box portion of the control will take on the content from the ComboBoxItem that had been selected.  Otherwise, it maintains its default text assigned during initialization.

git-svn-id: https://xamlite.svn.codeplex.com/svn@100911 81e6e737-5a20-4d0f-a97c-cb01297b46e0