Modified the RenderTransform variable for the XAMLiteImage class into an enum where the developer can select: FlipHorizontal, FlipVertical, FlipHorizontalAndVertical, Rotate90, Rotate180.
Made modify further to allow arbitrary angles.  

These are used to transform an image to cut down on the number of assets that may be required for building a special class, etc., such as a button where the right and left sides or top and bottom are mirror images of each other.  Or, in the case of the Scroll bar, the scroll bar can be transformed to be vertical or horizontal.

git-svn-id: https://xamlite.svn.codeplex.com/svn@101026 81e6e737-5a20-4d0f-a97c-cb01297b46e0